requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-experimental==0.0.49
openai==1.10.0
chromadb==0.4.18
sentence-transformers==2.2.2
networkx==3.2.1
matplotlib==3.8.2
plotly==5.17.0
pandas==2.1.4
numpy==1.24.3
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.28.1
pydantic==2.5.0
typing-extensions==4.8.0
tiktoken==0.5.2
faiss-cpu==1.7.4
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
neo4j==5.14.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl 